+++ backend/otel/otel.go
@@ -82,40 +82,6 @@ func projectToInt(projectID string) (int, error) {
 	return 0, e.New(fmt.Sprintf("invalid project id %s", projectID))
 }
 
-func getAttributesMap(resourceAttributes, eventAttributes map[string]any) map[string]string {
-	attributesMap := make(map[string]string)
-	for _, m := range []map[string]any{resourceAttributes, eventAttributes} {
-		for k, v := range m {
-			shouldSkip := false
-
-			for _, attr := range highlight.InternalAttributes {
-				if k == attr {
-					shouldSkip = true
-					continue
-				}
-			}
-
-			if shouldSkip {
-				continue
-			}
-
-			vStr := cast(v, "")
-			if vStr != "" {
-				attributesMap[k] = vStr
-			}
-			vInt := cast[int64](v, 0)
-			if vInt != 0 {
-				attributesMap[k] = strconv.FormatInt(vInt, 10)
-			}
-			vFlt := cast[float64](v, 0.)
-			if vFlt > 0. {
-				attributesMap[k] = strconv.FormatFloat(vFlt, 'f', -1, 64)
-			}
-		}
-	}
-	return attributesMap
-}
-
 func (o *Handler) HandleTrace(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	body, err := io.ReadAll(r.Body)
@@ -190,19 +156,17 @@ func (o *Handler) HandleTrace(w http.ResponseWriter, r *http.Request) {
 								log.WithContext(ctx).WithField("ProjectVerboseID", projectID).WithField("ExcMessage", excMessage).Errorf("otel span error got invalid project id")
 								return nil
 							}
-							attributesMap := getAttributesMap(resourceAttributes, eventAttributes)
 							logRow := clickhouse.NewLogRow(clickhouse.LogRowPrimaryAttrs{
 								Timestamp:       ts,
 								ProjectId:       uint32(projectIDInt),
 								TraceId:         traceID,
 								SpanId:          spanID,
 								SecureSessionId: sessionID,
-							})
+							}, clickhouse.WithLogAttributes(resourceAttributes, eventAttributes))
 							logRow.SeverityText = "ERROR"
 							logRow.SeverityNumber = int32(log.ErrorLevel)
 							logRow.ServiceName = serviceName
 							logRow.Body = excMessage
-							logRow.LogAttributes = attributesMap
 							if projectID != "" {
 								if _, ok := projectLogs[projectID]; !ok {
 									projectLogs[projectID] = []*clickhouse.LogRow{}
@@ -274,7 +238,6 @@ func (o *Handler) HandleTrace(w http.ResponseWriter, r *http.Request) {
 							log.WithContext(ctx).WithField("ProjectVerboseID", projectID).WithField("LogMessage", logMessage).Errorf("otel span log got invalid project id")
 							continue
 						}
-						attributesMap := getAttributesMap(resourceAttributes, eventAttributes)
 						lvl, _ := log.ParseLevel(logSev)
 						logRow := clickhouse.NewLogRow(clickhouse.LogRowPrimaryAttrs{
 							Timestamp:       event.Timestamp().AsTime(),
@@ -282,12 +245,11 @@ func (o *Handler) HandleTrace(w http.ResponseWriter, r *http.Request) {
 							SpanId:          span.SpanID().String(),
 							ProjectId:       uint32(projectIDInt),
 							SecureSessionId: sessionID,
-						})
+						}, clickhouse.WithLogAttributes(resourceAttributes, eventAttributes))
 						logRow.SeverityText = logSev
 						logRow.SeverityNumber = int32(lvl)
 						logRow.ServiceName = serviceName
 						logRow.Body = logMessage
-						logRow.LogAttributes = attributesMap
 						if projectID != "" {
 							if _, ok := projectLogs[projectID]; !ok {
 								projectLogs[projectID] = []*clickhouse.LogRow{}
@@ -435,19 +397,17 @@ func (o *Handler) HandleLog(w http.ResponseWriter, r *http.Request) {
 					log.WithContext(ctx).WithField("ProjectID", projectID).WithField("LogMessage", logRecord.Body().AsRaw()).Errorf("otel log got invalid project id")
 					continue
 				}
-				attributesMap := getAttributesMap(resourceAttributes, logAttributes)
 				logRow := clickhouse.NewLogRow(clickhouse.LogRowPrimaryAttrs{
 					Timestamp:       logRecord.Timestamp().AsTime(),
 					TraceId:         logRecord.TraceID().String(),
 					SpanId:          logRecord.SpanID().String(),
 					ProjectId:       uint32(projectIDInt),
 					SecureSessionId: sessionID,
-				})
+				}, clickhouse.WithLogAttributes(resourceAttributes, logAttributes))
 				logRow.SeverityText = logRecord.SeverityText()
 				logRow.SeverityNumber = int32(logRecord.SeverityNumber())
 				logRow.ServiceName = serviceName
 				logRow.Body = logRecord.Body().Str()
-				logRow.LogAttributes = attributesMap
 				if projectID != "" {
 					if _, ok := projectLogs[projectID]; !ok {
 						projectLogs[projectID] = []*clickhouse.LogRow{}
